add set-up, do not need download

+++ set_up.sh
@@ -1,8 +1,14 @@
 #!/usr/bin/env bash
 
-docker run --rm -v `pwd`:`pwd` -w `pwd` -i limanling/uiuc_ie_m18 mkdir -p ${PWD}/system/aida_edl
-docker run -v ${PWD}/system/aida_edl:/data panx27/data-processor wget http://159.89.180.81/demo/resources/edl_data.tar.gz -P /data
-docker run -v ${PWD}/system/aida_edl:/data panx27/data-processor tar zxvf /data/edl_data.tar.gz -C /data
+if [ -d "${PWD}/system/aida_edl" ]
+then
+    echo "KB for linking is already in "${PWD}"/system/aida_edl"
+else
+    docker run --rm -v `pwd`:`pwd` -w `pwd` -i limanling/uiuc_ie_m18 mkdir -p ${PWD}/system/aida_edl
+    docker run -v ${PWD}/system/aida_edl:/data panx27/data-processor wget http://159.89.180.81/demo/resources/edl_data.tar.gz -P /data
+    docker run -v ${PWD}/system/aida_edl:/data panx27/data-processor tar zxvf /data/edl_data.tar.gz -C /data
+
+fi
 docker run --rm -v ${PWD}/system/aida_edl/edl_data/db:/data/db --name db mongo
 
 docker run -d -i -t --rm -w /aida_nominal_coreference_en -p 2468:2468 --name nominal_coreference wangqy96/aida_nominal_coreference_en python nominal_backend.py
